+++ lib/entities/index.ts
@@ -1,0 +1,4 @@
+export * from "./entity.types.ts";
+export * from "./entity-registry.ts";
+export * from "./entity-builder.ts";
+export * from "./entity.service.ts";