+++ lib/products/index.ts
@@ -2,3 +2,4 @@ export * from "./product-builder";
 export * from "./product-cache";
 export * from "./product.service";
 export * from "./product.types";
+export * from "./product-entity-adapter";